Use today as transfer date for #10 (and valnet #510)
As opposed to commit.commitment_date; commit is remaining unassigned, so that throws.  All of the Commitment creation mechanisms available here use the default commitment date, which is datetime.date.today().  The commit being None is not good on its own, but avoiding the unexpected throw will let me get to the bottom of it.